- Renamed methods - Added setup.py - Added init.py files - Added some prints

+++ simulator/provider/IProvider.py
@@ -4,7 +4,6 @@ from typing import Dict, Any
 from simulator.messages.CommunicationCommand import CommunicationCommand
 from simulator.messages.MobilityCommand import MobilityCommand
 
-
 class IProvider(ABC):
     @abstractmethod
     def send_communication_command(self, command: CommunicationCommand):
@@ -19,7 +18,7 @@ class IProvider(ABC):
         pass
 
     @abstractmethod
-    def current_time(self) -> int:
+    def current_time(self) -> float:
         pass
 
     tracked_variables: Dict[str, Any]
